refactor(audit): move activation into runtime facade
