+++ crates/audit/src/pipeline.rs
@@ -12,9 +12,13 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 
-use crate::{AuditEntry, AuditResult, observability, system::AuditTargetMetricSnapshot};
+use crate::{
+    AuditEntry, AuditResult, observability,
+    system::{AuditSystemState, AuditTargetMetricSnapshot},
+};
 use rustfs_targets::{Target, target::EntityTarget};
 use std::sync::Arc;
+use std::time::Duration;
 use tokio::sync::{Mutex, RwLock};
 use tracing::{error, info, warn};
 
@@ -289,6 +293,95 @@ impl AuditRuntimeFacade {
         drop(registry);
         *self.replay_workers.write().await = activation.replay_workers;
         Ok(())
+    }
+
+    pub async fn activate_targets_with_replay(
+        &self,
+        state: Arc<RwLock<AuditSystemState>>,
+        targets: Vec<Box<dyn Target<AuditEntry> + Send + Sync>>,
+    ) -> rustfs_targets::RuntimeActivation<AuditEntry> {
+        rustfs_targets::activate_targets_with_replay(targets, |target| {
+            let state = state.clone();
+            async move {
+                rustfs_targets::init_target_and_optionally_start_replay(
+                    target,
+                    |target_id, has_replay| {
+                        if has_replay {
+                            info!(target_id = %target_id, "Audit stream processing started");
+                        } else {
+                            info!(target_id = %target_id, "No store configured, skip audit stream processing");
+                        }
+                    },
+                    move |store, target| {
+                        rustfs_targets::start_replay_worker(
+                            store,
+                            target,
+                            Arc::new(move |event| {
+                                let state = state.clone();
+                                Box::pin(async move {
+                                    if !matches!(
+                                        *state.read().await,
+                                        AuditSystemState::Running | AuditSystemState::Paused | AuditSystemState::Starting
+                                    ) {
+                                        return;
+                                    }
+                                    match event {
+                                        rustfs_targets::ReplayEvent::Delivered { key, target } => {
+                                            info!(
+                                                "Successfully sent audit entry, target: {}, key: {}",
+                                                target.id(),
+                                                key.to_string()
+                                            );
+                                            observability::record_target_success();
+                                        }
+                                        rustfs_targets::ReplayEvent::RetryableError { error, target, .. } => match error {
+                                            rustfs_targets::TargetError::NotConnected => {
+                                                warn!("Target {} not connected, retrying...", target.id());
+                                            }
+                                            rustfs_targets::TargetError::Timeout(_) => {
+                                                warn!("Timeout sending to target {}, retrying...", target.id());
+                                            }
+                                            _ => {}
+                                        },
+                                        rustfs_targets::ReplayEvent::Dropped { reason, target, .. } => {
+                                            warn!("Dropped queued payload for target {}: {}", target.id(), reason);
+                                            observability::record_target_failure();
+                                        }
+                                        rustfs_targets::ReplayEvent::PermanentFailure { error, target, .. } => {
+                                            error!("Permanent error for target {}: {}", target.id(), error);
+                                            target.record_final_failure();
+                                            observability::record_target_failure();
+                                        }
+                                        rustfs_targets::ReplayEvent::RetryExhausted { key, target } => {
+                                            warn!(
+                                                "Max retries exceeded for key {}, target: {}, skipping",
+                                                key.to_string(),
+                                                target.id()
+                                            );
+                                            target.record_final_failure();
+                                            observability::record_target_failure();
+                                        }
+                                        rustfs_targets::ReplayEvent::UnreadableEntry { key, error, target } => {
+                                            warn!(
+                                                "Skipping unreadable audit store entry {} for target {}: {}",
+                                                key,
+                                                target.id(),
+                                                error
+                                            );
+                                        }
+                                    }
+                                })
+                            }),
+                            None,
+                            Duration::from_millis(500),
+                            Duration::from_millis(500),
+                        )
+                    },
+                )
+                .await
+            }
+        })
+        .await
     }
 
     pub async fn stop_replay_workers(&self) {
